SEQNG-468 Refactoring
